requirements.txt
streamlit==1.12.0
pandas==2.2.1
google-cloud-storage==2.15.0
joblib==1.3.2